+++ calculator/src/test/java/calculator/TestOpStack.java
@@ -2,6 +2,7 @@ package calculator;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
+
 import java.util.EmptyStackException;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -56,6 +57,12 @@ class TestOpStack {
         "After stack has been popped it should be empty.");
   }
 
-
+  @Test
+  void testOpStackPeek() {
+    opStack.push(Symbol.DIVIDE);
+    assertEquals("/", opStack.peek(), "Pushed divide symbol so should be on top when peeking.");
+    assertEquals(opStack.isEmpty(), false,
+        "Peek should not remove element from stack.");
+  }
 
 }